fix(analytics): set duration field to precise decimal as episode's audio file duration

+++ app/Libraries/Analytics/Database/Migrations/2017-12-01-210000_add_analytics_podcasts_procedure.php
@@ -36,7 +36,7 @@ class AddAnalyticsPodcastsProcedure extends Migration
             IN `p_os` VARCHAR(32),
             IN `p_bot` TINYINT(1) UNSIGNED,
             IN `p_filesize` INT UNSIGNED,
-            IN `p_duration` INT UNSIGNED,
+            IN `p_duration` DECIMAL(8,3) UNSIGNED,
             IN `p_age` INT UNSIGNED,
             IN `p_new_listener` TINYINT(1) UNSIGNED
             )  MODIFIES SQL DATA
